requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# requirements.txt
# AI Tuberculosis Detection and Severity Assessment System
# ─────────────────────────────────────────────────────────────────────────────

# Core deep learning framework (PyTorch backend)
torch>=2.0.0
torchvision>=0.15.0

# Ultralytics YOLO — Detection model framework
ultralytics>=8.0.0

# Computer vision — Image loading, augmentation, drawing
opencv-python>=4.8.0

# Numerical computing
numpy>=1.24.0

# Plotting and visualization — confusion matrices, training curves
matplotlib>=3.7.0

# Machine learning metrics — precision, recall, confusion matrix
scikit-learn>=1.3.0

# Progress bars for training/inference loops
tqdm>=4.65.0

# Image processing utilities
Pillow>=9.5.0

# YAML parsing — for dataset configuration files
PyYAML>=6.0

# Roboflow Python SDK — Easy dataset download from Roboflow
roboflow>=1.1.0

# Streamlit web interface
streamlit>=1.30.0
